refactor(agent-ui): move workspace components to settings, remove orphaned CSS
Move 10 components (WorkspaceTree, WorkspaceNode, WorktreeNode,
WorktreeDetailPopover, WorktreeDetailContent, TreeOverflowMenu,
AddWorktreeDialog, RemoveWorktreeDialog, RemoveWorkspaceDialog,
PhraseListField) and useTreeKeyboard hook from session to settings entry.

Extract .ws-*, .settings-label, .agent-add-*, .agent-badge CSS to
settings index.css. Remove orphaned .hud-settings-*, .agent-dir-*,
.agent-pw-*, .ws-* CSS from session. Clean up useAgentSettings to
only expose loading, settings, setSettings, setAndSaveSettings, apiKey.
