+++ mods/agent/ui/settings/src/components/AddWorktreeDialog.tsx
