+++ mods/agent/ui/settings/src/components/RemoveWorkspaceDialog.tsx
