+++ mods/agent/ui/settings/src/components/RemoveWorktreeDialog.tsx
