+++ mods/agent/ui/settings/src/components/WorkspaceNode.tsx
