+++ mods/agent/ui/settings/src/components/WorktreeDetailContent.tsx
