+++ mods/agent/ui/settings/src/components/WorktreeDetailPopover.tsx
