+++ mods/agent/ui/settings/src/components/WorktreeNode.tsx
