+++ mods/agent/ui/settings/src/hooks/useTreeKeyboard.ts
